Revert "Ajout de la possibilité de choisir son lycée "
