--- mp2i/static/text/Liste_lycee_MP2I.txt
@@ -1,40 +1,0 @@
-Alain René Lesage
-Alfred Kastler
-Aux Lazaristes
-Berthollet
-Camille Guérin
-Carnot
-Centre International de Valbonne
-Champollion
-Charles Coëffin
-Claude Bernard 
-Claude Fauriel
-Clémenceau
-Colbert
-Descartes
-Faidherbe
-Fénelon Sainte-Marie
-Fermat
-Franklin Roosevelt
-Frédéric Ozanam
-Gay-Lussac
-Guy Mollet
-Henri Poincaré
-Henri Wallon
-Hoche
-Janson de Sailly
-Jean XXIII 
-Joffre
-Kléber
-La Fayette
-La Martinière Monplaisir
-Le Parc
-Leconte de Lisle
-Louis Le Grand
-Louis Thuillier
-Montaigne
-Paul Valéry
-Pierre Corneille
-Saint-Louis
-Thiers
-Victor Hugo
