+++ mp2i/wrappers/member.py
@@ -78,7 +78,6 @@ class MemberWrapper:
                 guild_id=self.guild.id,
                 name=self.member.name,
                 role=qualifier,
-                school=None,
             )
         )
         self.__model = self._fetch()  # Update the model
@@ -106,11 +105,3 @@ class MemberWrapper:
     @profile_color.setter
     def profile_color(self, value: str):
         self.update(profile_color=value)
-
-    @property
-    def school(self) -> str:
-        return self.__model.school or "Aucun"
-
-    @school.setter
-    def school(self, value: str):
-        return self.update(school=value)
